+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+##### Fixed
+* **\[core\]** Ensure that `@ObservableRef` can be used for `@Observable` properties with a generic type.
+
 ### [v0.32](https://github.com/arez/arez/tree/v0.32) (2017-12-01)
 [Full Changelog](https://github.com/arez/arez/compare/v0.31...v0.32)